AES.API

Documentation

keygenIO :: Int -> String -> IO ()

Takes an Int size and a String filename. Generates a key based on keygen and writes the key to the file.

encryptIO :: String -> String -> String -> IO ()

Takes 3 Strings, keyFile, fileIn, fileOut. fileIn is encrypted based on encrypt with keyFile, and the ciphertext is written to fileOut.

decryptIO :: String -> String -> String -> IO ()

Takes 3 Strings, keyFile, fileIn, fileOut. fileIn is decrypted based on decrypt with keyFile, and the decrypted text is written to fileOut.

keygen :: RandomGen a => a -> Int -> Maybe ByteString

Takes a RandomGen g, Int size. If size is one of 128, 192, or 256 then generates a number n in [2^(size-1), 2^size - 1] and returns Just ByteString based on n. Otherwise returns Nothing.

encrypt :: ByteString -> ByteString -> ByteString

Takes in two ByteStrings, k, m representing the key and message to be encrypted. Returns the encrypted ByteString.

decrypt :: ByteString -> ByteString -> ByteString

Takes in two ByteStrings, k, c representing the key and ciphertext to be dcecrypted. Returns the decrypted ByteString.