AES.Internal

Description

Internal functions used for AES.

Documentation

ecb :: Int -> (Int -> [[Word8]] -> [[Word8]] -> [[Word8]]) -> [[Word8]] -> [Word8] -> ByteString -> ByteString

Takes an Int maxRound, encryption function f, [[Word8]] roundKeys, [Word8] message, and starting matrix []. Encrypts each block with f and concatenates all blocks together.

setupForTransform :: ByteString -> (Int, [[Word8]])

For ByteString key, returns (maxRound, roundKeys).

encryptInit :: Int -> [[Word8]] -> [[Word8]] -> [[Word8]]

Round 0 Takes an Int maxRound, [[Word8]] roundKeys, [Word8] state, and returns the modified state.

decryptInit :: Int -> [[Word8]] -> [[Word8]] -> [[Word8]]

Round Nr Takes an Int maxRound, [[Word8]] roundKeys, [Word8] state, and returns the modified state.

rotate :: [a] -> [a]

Rotates a list by 1.

rotate [1,2,3] -> [2,3,1]

fieldMult :: Word8 -> Word8 -> Word8

We only need to implement multiplication by 2 and 3 in GF(2^8) for encryption, 9, 11, 13, and 14 for decryption. Multiplication by 2 is equivalent to bit shifting by one and adding (xor) 0x1b (27) if the high bit was set. 3 x b = (2 xor 1) b = (2 x b) xor b