Common

Description

Exports utility functions.

Documentation

padToN :: Int -> [Word8] -> (Word8, [Word8])

Takes an Int n and a Word8 list w and returns a tuple (a, [b]) where [b] is w padded with zeroes to the nearest greater multiple of n and a is the amount of padding needed.

Example:

  padToN 3 [23,12,64,32] -> (2,[23,12,64,32,0,0])
  padToN 3 [23,12,64,32,6,42] -> (0,[23,12,64,32,6,42])

intToWord8List :: Integer -> [Word8] -> [Word8]

Converts an Integer into a Word8 list where each element in the list represents a byte. The inverse of word8ListToInt.

Example:

  intToWord8List 42310 -> [165, 70], or 0xA546.
  intToWord8List 52582001236900766738172 -> [11,34,121,219,13,148,44,170,230,252], or 0xB2279DB0D942CAAE6FC.

word8ListToInt :: [Word8] -> Integer

Converts a Word8 list of bytes into an Integer. The inverse of intToWord8List.

Example:

word8ListToInt [165,70] -> 42310
word8ListToInt [11,34,121,219,13,148,44,170,230,252] -> 52582001236900766738172

xorByte :: Bits a => [a] -> a -> [a]

Xors every element in a Bit list with the parameter Bit.

Example:

xorByte [b1, b2 ... bn] b -> [b1 ⊕ b, b2 ⊕ b ... bn ⊕ b]

xorList :: Bits a => [a] -> [a] -> [a]

For two Bit lists, returns a new list with each corresponding Bit xor'd.

Example:

xorList [b1, b2 ... bn] [c1, c2 ... cn] -> [b1 ⊕ c1, b2 ⊕ c2 ... bn ⊕ cn]

xorMatrix :: Bits a => [[a]] -> [[a]] -> [[a]]

For two Bit matrices, returns a new matrix with each corresponding Bit xor'd.

Example:

xorMatrix [[b11, b12] [b12, b22]] [[c11, c12] [c12, c22]] -> [[b11 ⊕ c11, b12 ⊕ c12] [b12 ⊕ c12, b22 ⊕ c22]]

flatListToMatrix :: Int -> [a] -> [[a]] -> [[a]]

For Integer n, list xs, and starting matrix m, transforms m into an n x n matrix representation for xs.

Example:

flatListToMatrix 4 [b1 ... b16] -> [[b1 ... b4], [b5 ... b8], [b9 ... b12], [b13 ... n16]]

byteStringToInt :: ByteString -> (Integer, ByteString)

For a ByteString lb, where l is the two-byte representation of the integer length of the next integer in b, returns (i, rem) where i is integer representation of the next l bytes in b, and rem is the remaining bytes, if any.

Example:

byteStringToInt (pack [0,4,234,66,25,181,43,23]) -> (3930200501, pack [43,23]) (l = 04, b = [234,66,25,181] or 0xEA4219B5)

intToByteString :: Integer -> ByteString

Turns an Integer into ByteString lb where b is n's ByteString representation and l is its two-byte length.

Example:

intToByteString 3930200501 -> pack [0,4,234,66,25,181]