NumberTheory

Description

Exports functions for generating prime numbers based on Miller-Rabin primality test.

Documentation

genPrime :: (RandomGen a, Integral b) => a -> b -> Integer

Generates a random prime with b bits using the Miller-Rabin primality test.

powModN :: Integral a => a -> a -> a -> a

For a, d, n, returns \( a^d \pmod n \).

findInverse :: Integral a => a -> a -> a -> a

For a, p, q, finds the multiplicative inverse of a in \(\mathbb{Z}/pq\mathbb{Z}\). That is, finds e such that \[ ae \equiv 1 \pmod{\varphi(pq)} \].