RSA.API

Documentation

keygenIO :: Integral a => a -> String -> String -> IO ()

Takes an Integral size and two strings pubFileOut, prvFileOut. Generates a public/private key pair based on keygen and writes the key pair to the files.

encryptIO :: String -> String -> String -> IO ()

Takes 3 Strings, keyFile, fileIn, fileOut. fileIn is encrypted based on encrypt with keyFile, and the ciphertext is written to fileOut.

decryptIO :: String -> String -> String -> IO ()

Takes 3 Strings, keyFile, fileIn, fileOut. fileIn is decrypted based on decrypt with keyFile, and the decrypted text is written to fileOut.

keygen :: (RandomGen a, Integral b) => a -> b -> Maybe (ByteString, ByteString)

Takes a RandomGen g, Integral size. If is size one of 1024, 2048, or 4096 then returns Just (ByteString, ByteString), where the ByteStrings are the public and private keys, respectively. If size is anything then Nothing is returned.

encrypt :: RandomGen a => a -> ByteString -> ByteString -> ByteString

Takes a RandomGen g and two ByteStrings k, m, representing the key and message to be encrypted. m is encrypted with AES-256 in ECB mode (I know) to produce ciphertext c. The generated AES key l is then encrypted with the public key k, so the final output is RSA_k(l)|AES_l(m).

decrypt :: ByteString -> ByteString -> ByteString

Takes two ByteStrings k, m, representing the key and ciphertext to be decrypted.