RSA.API

Description

This is the API for the main RSA functions. Supports 1024, 2048, and 4096 bit keys. This is for file encryption, so it encrypts the file with AES-256 then encrypts the AES key using RSA. For RSA public key pk, generated AES key k, and message m, the ciphertext is:

(Enc_RSA_pk(k)|Enc_AES_k(m))

Documentation

keygenIO :: Integral a => a -> String -> String -> IO ()

For Integral n, String pkFileOut, String skFileOut, creates an n bit (pk, sk) key pair and writes the keys to pkFileOut and skFileOut, respectively.

encryptIO :: String -> String -> String -> IO ()

For Strings keyFile, fileIn, fileOut, uses the key in keyFile to encrypt the contents of fileIn, writing the ciphertext to fileOut.

decryptIO :: String -> String -> String -> IO ()

For Strings keyFile, fileIn, fileOut, uses the key in keyFile to decrypt the contents of fileIn, writing the message to fileOut.

keygen :: (RandomGen g, Integral a) => g -> a -> Maybe (ByteString, ByteString)

For RandomGen g and Integral n in (1024, 2048, 4096), returns Just (pk, sk). If n is not a valid key-size then returns Nothing.

encrypt :: RandomGen g => g -> ByteString -> ByteString -> ByteString

For RandomGen g and ByteStrings k and m, returns the encrypted ByteString ciphertext. If AES keygen fails then returns empty.

decrypt :: ByteString -> ByteString -> ByteString

For ByteStrings k and c, returns the decrypted ByteString message.